+++ packages/ngx-deploy-npm/src/executors/deploy/schema.d.ts
@@ -1,5 +1,9 @@
 export interface DeployExecutorOptions {
   /**
+   * Indicate the dist folder path. This is useful when ngx-deploy-npm can not detect your library dist folder automatically. The path should be relative to the project's root
+   */
+  distFolderPath?: string;
+  /**
    * A named build target, as specified in the `configurations` section of workspace/angular.json. Each named target is accompanied by a configuration of option defaults for that target. This is equivalent to calling the command `[nx|ng] build --configuration=XXX`.
    */
   buildTarget?: string;
@@ -18,11 +22,11 @@ export interface DeployExecutorOptions {
   /**
    * Tells the registry whether this package should be published as public or restricted. Only applies to scoped packages, which default to restricted. If you don’t have a paid account, you must publish with --access public to publish scoped packages.
    */
-  access?: string;
+  access?: 'public' | 'restricted';
   /**
    * If you have two-factor authentication enabled in auth-and-writes mode then you can provide a code from your authenticator with this. If you don’t include this and you’re running from a TTY then you’ll be prompted.
    */
-  otp?: string;
+  otp?: string | number;
   /**
    * For testing: Run through without making any changes. Execute with --dry-run and nothing will happen.
    */