+++ packages/ngx-deploy-npm/src/executors/deploy/utils/index.ts
@@ -1,6 +1,6 @@
 export * from './default-options';
 export * from './execute-async';
-export * from './fs-async';
+export * from './file-utils';
 export * from './get-lib-output-path';
 export * from './interfaces';
 export * from './set-default-options';